archymedes_bk_challenge/requirements.txt
apispec==4.4.0
asgiref==3.4.1
attrs==20.3.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
click==7.1.2
Django==3.2.9
django-cors-headers==3.10.0
django-environ==0.8.1
django-filter==21.1
django-rest-passwordreset==1.2.1
djangorestframework==3.12.4
ecdsa==0.14.1
flasgger==0.9.5
Flask==1.1.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
idna==3.3
importlib-metadata==3.10.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.11.1
mistune==0.8.4
mysql-connector-python==8.0.27
protobuf==3.19.1
psycopg2==2.9.1
pyasn1==0.4.8
PyJWT==2.3.0
PyMySQL==1.0.2
pyrsistent==0.17.3
python-dotenv==0.17.0
python-http-client==3.3.3
python-jose==3.2.0
python-keycloak==0.24.0
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
sendgrid==6.9.0
six==1.15.0
sqlparse==0.4.2
starkbank-ecdsa==1.1.1
twilio==7.3.0
typing-extensions==3.7.4.3
urllib3==1.26.7
Werkzeug==1.0.1
zipp==3.4.1
